+++ perf/flame.jl
@@ -40,7 +40,7 @@ ProfileCanvas.html_file(joinpath(output_dir, "flame.html"), results)
 allocs_limit = Dict()
 allocs_limit["flame_baroclinic_wave_moist_gpu"] = 1_243_048
 allocs_limit["flame_default"] = 99_928
-allocs_limit["flame_default_1m"] = 596_096
+allocs_limit["flame_default_1m"] = 663_056
 allocs_limit["flame_diagnostics"] = 10_677_144
 allocs_limit["flame_aquaplanet_diagedmf"] = 11_644_128
 allocs_limit["flame_aquaplanet_progedmf"] = 774_712
@@ -107,8 +107,4 @@ end
               allocs ≤
               allocs_limit[job_id] * buffer
     end
-end
-
-if config.comms_ctx isa ClimaComms.SingletonCommsContext && !isinteractive()
-    include(joinpath(pkgdir(CA), "perf", "jet_report_nfailures.jl"))
 end