+++ src/cache/cache.jl
@@ -88,20 +88,20 @@ function build_cache(
     atmos,
     params,
     surface_setup,
-    sim_info,
+    dt,
+    start_date,
     aerosol_names,
     time_varying_trace_gas_names,
     steady_state_velocity,
 )
-    (; dt, start_date, output_dir) = sim_info
     FT = eltype(params)
     dt = FT(dt)
 
     ᶜcoord = Fields.local_geometry_field(Y.c).coordinates
     ᶠcoord = Fields.local_geometry_field(Y.f).coordinates
     grav = FT(CAP.grav(params))
-    ᶜΦ = grav .* ᶜcoord.z
-    ᶠΦ = grav .* ᶠcoord.z
+    ᶜΦ = geopotential.(grav, ᶜcoord.z)
+    ᶠΦ = geopotential.(grav, ᶠcoord.z)
 
     (; ᶜf³, ᶠf¹²) = compute_coriolis(ᶜcoord, ᶠcoord, params)
 
@@ -174,7 +174,6 @@ function build_cache(
             time_varying_trace_gas_names,
             atmos.insolation,
         ) : ()
-
     hyperdiff = hyperdiffusion_cache(Y, atmos)
     non_orographic_gravity_wave = non_orographic_gravity_wave_cache(Y, atmos)
     orographic_gravity_wave = orographic_gravity_wave_cache(Y, atmos)
@@ -212,7 +211,6 @@ function compute_coriolis(ᶜcoord, ᶠcoord, params)
         Ω = CAP.Omega(params)
         global_geom = Spaces.global_geometry(axes(ᶜcoord))
         if global_geom isa Geometry.DeepSphericalGlobalGeometry
-            @info "using deep atmosphere"
             coriolis_deep(coord::Geometry.LatLongZPoint) = Geometry.LocalVector(
                 Geometry.Cartesian123Vector(zero(Ω), zero(Ω), 2 * Ω),
                 global_geom,