+++ src/prognostic_equations/advection.jl
@@ -18,7 +18,7 @@ Specifically, this function calculates:
 - Horizontal advection of EDMFX updraft density-area product (`ρaʲ`).
 - Horizontal advection of total energy (`ρe_tot`) using total enthalpy flux.
 - Horizontal advection of EDMFX updraft moist static energy (`mseʲ`).
-- Horizontal advection of turbulent kinetic energy (`ρatke⁰`) if used.
+- Horizontal advection of turbulent kinetic energy (`ρtke`) if used.
 - Horizontal pressure gradient, kinetic energy gradient, and geopotential gradient
   forces for horizontal momentum (`uₕ`).
 
@@ -31,7 +31,7 @@ Arguments:
 - `t`: Current simulation time (not directly used in calculations).
 
 Modifies `Yₜ.c.ρ`, `Yₜ.c.ρe_tot`, `Yₜ.c.uₕ`, and EDMFX-related fields in
-`Yₜ.c.sgsʲs` and `Yₜ.c.sgs⁰` if applicable.
+`Yₜ.c.sgsʲs` and `Yₜ.c.ρtke` if applicable.
 """
 NVTX.@annotate function horizontal_dynamics_tendency!(Yₜ, Y, p, t)
     n = n_mass_flux_subdomains(p.atmos.turbconv_model)
@@ -45,35 +45,46 @@ NVTX.@annotate function horizontal_dynamics_tendency!(Yₜ, Y, p, t)
         (; ᶜuʲs) = p.precomputed
     end
 
-    @. Yₜ.c.ρ -= wdivₕ(Y.c.ρ * ᶜu)
+    @. Yₜ.c.ρ -= split_divₕ(Y.c.ρ * ᶜu, 1)
     if p.atmos.turbconv_model isa PrognosticEDMFX
         for j in 1:n
-            @. Yₜ.c.sgsʲs.:($$j).ρa -= wdivₕ(Y.c.sgsʲs.:($$j).ρa * ᶜuʲs.:($$j))
+            @. Yₜ.c.sgsʲs.:($$j).ρa -= split_divₕ(
+                Y.c.sgsʲs.:($$j).ρa * ᶜuʲs.:($$j),
+                1,
+            )
         end
     end
 
     ᶜe_tot = @. lazy(specific(Y.c.ρe_tot, Y.c.ρ))
     ᶜh_tot = @. lazy(TD.total_specific_enthalpy(thermo_params, ᶜts, ᶜe_tot))
-    @. Yₜ.c.ρe_tot -= wdivₕ(Y.c.ρ * ᶜh_tot * ᶜu)
+    @. Yₜ.c.ρe_tot -= split_divₕ(Y.c.ρ * ᶜu, ᶜh_tot)
 
     if p.atmos.turbconv_model isa PrognosticEDMFX
         for j in 1:n
             @. Yₜ.c.sgsʲs.:($$j).mse -=
-                wdivₕ(Y.c.sgsʲs.:($$j).mse * ᶜuʲs.:($$j)) -
-                Y.c.sgsʲs.:($$j).mse * wdivₕ(ᶜuʲs.:($$j))
+                split_divₕ(ᶜuʲs.:($$j), Y.c.sgsʲs.:($$j).mse) -
+                Y.c.sgsʲs.:($$j).mse * split_divₕ(ᶜuʲs.:($$j), 1)
         end
     end
 
     if use_prognostic_tke(p.atmos.turbconv_model)
-        @. Yₜ.c.sgs⁰.ρatke -= wdivₕ(Y.c.sgs⁰.ρatke * ᶜu)
+        ᶜtke = @. lazy(specific(Y.c.ρtke, Y.c.ρ))
+        @. Yₜ.c.ρtke -= split_divₕ(Y.c.ρ * ᶜu, ᶜtke)
     end
 
-    # This is equivalent to grad_h(Φ + K) + grad_h(p) / ρ
     ᶜΦ_r = @. lazy(phi_r(thermo_params, ᶜts))
     ᶜθ_v = @. lazy(theta_v(thermo_params, ᶜts))
     ᶜθ_vr = @. lazy(theta_vr(thermo_params, ᶜts))
     ᶜΠ = @. lazy(dry_exner_function(thermo_params, ᶜts))
-    @. Yₜ.c.uₕ -= C12(gradₕ(ᶜK + ᶜΦ - ᶜΦ_r) + cp_d * (ᶜθ_v - ᶜθ_vr) * gradₕ(ᶜΠ))
+    ᶜθ_v_diff = @. lazy(ᶜθ_v - ᶜθ_vr)
+    # split form pressure gradient: 0.5 * cp_d * [θv ∇Π + ∇(θv Π) - Π∇θv]
+    @. Yₜ.c.uₕ -= C12(
+        gradₕ(ᶜK + ᶜΦ - ᶜΦ_r) +
+        cp_d *
+        (
+            ᶜθ_v_diff * gradₕ(ᶜΠ) + gradₕ(ᶜθ_v_diff * ᶜΠ) - ᶜΠ * gradₕ(ᶜθ_v_diff)
+        ) / 2,
+    )
     # Without the C12(), the right-hand side would be a C1 or C2 in 2D space.
     return nothing
 end
@@ -110,39 +121,40 @@ NVTX.@annotate function horizontal_tracer_advection_tendency!(Yₜ, Y, p, t)
     end
 
     for ρχ_name in filter(is_tracer_var, propertynames(Y.c))
-        @. Yₜ.c.:($$ρχ_name) -= wdivₕ(Y.c.:($$ρχ_name) * ᶜu)
+        ᶜχ = @. lazy(specific(Y.c.:($$ρχ_name), Y.c.ρ))
+        @. Yₜ.c.:($$ρχ_name) -= split_divₕ(Y.c.ρ * ᶜu, ᶜχ)
     end
 
     if p.atmos.turbconv_model isa PrognosticEDMFX
         for j in 1:n
             @. Yₜ.c.sgsʲs.:($$j).q_tot -=
-                wdivₕ(Y.c.sgsʲs.:($$j).q_tot * ᶜuʲs.:($$j)) -
-                Y.c.sgsʲs.:($$j).q_tot * wdivₕ(ᶜuʲs.:($$j))
+                split_divₕ(ᶜuʲs.:($$j), Y.c.sgsʲs.:($$j).q_tot) -
+                Y.c.sgsʲs.:($$j).q_tot * split_divₕ(ᶜuʲs.:($$j), 1)
             if p.atmos.moisture_model isa NonEquilMoistModel && (
                 p.atmos.microphysics_model isa Microphysics1Moment ||
                 p.atmos.microphysics_model isa Microphysics2Moment
             )
                 @. Yₜ.c.sgsʲs.:($$j).q_liq -=
-                    wdivₕ(Y.c.sgsʲs.:($$j).q_liq * ᶜuʲs.:($$j)) -
-                    Y.c.sgsʲs.:($$j).q_liq * wdivₕ(ᶜuʲs.:($$j))
+                    split_divₕ(ᶜuʲs.:($$j), Y.c.sgsʲs.:($$j).q_liq) -
+                    Y.c.sgsʲs.:($$j).q_liq * split_divₕ(ᶜuʲs.:($$j), 1)
                 @. Yₜ.c.sgsʲs.:($$j).q_ice -=
-                    wdivₕ(Y.c.sgsʲs.:($$j).q_ice * ᶜuʲs.:($$j)) -
-                    Y.c.sgsʲs.:($$j).q_ice * wdivₕ(ᶜuʲs.:($$j))
+                    split_divₕ(ᶜuʲs.:($$j), Y.c.sgsʲs.:($$j).q_ice) -
+                    Y.c.sgsʲs.:($$j).q_ice * split_divₕ(ᶜuʲs.:($$j), 1)
                 @. Yₜ.c.sgsʲs.:($$j).q_rai -=
-                    wdivₕ(Y.c.sgsʲs.:($$j).q_rai * ᶜuʲs.:($$j)) -
-                    Y.c.sgsʲs.:($$j).q_rai * wdivₕ(ᶜuʲs.:($$j))
+                    split_divₕ(ᶜuʲs.:($$j), Y.c.sgsʲs.:($$j).q_rai) -
+                    Y.c.sgsʲs.:($$j).q_rai * split_divₕ(ᶜuʲs.:($$j), 1)
                 @. Yₜ.c.sgsʲs.:($$j).q_sno -=
-                    wdivₕ(Y.c.sgsʲs.:($$j).q_sno * ᶜuʲs.:($$j)) -
-                    Y.c.sgsʲs.:($$j).q_sno * wdivₕ(ᶜuʲs.:($$j))
+                    split_divₕ(ᶜuʲs.:($$j), Y.c.sgsʲs.:($$j).q_sno) -
+                    Y.c.sgsʲs.:($$j).q_sno * split_divₕ(ᶜuʲs.:($$j), 1)
             end
             if p.atmos.moisture_model isa NonEquilMoistModel &&
                p.atmos.microphysics_model isa Microphysics2Moment
                 @. Yₜ.c.sgsʲs.:($$j).n_liq -=
-                    wdivₕ(Y.c.sgsʲs.:($$j).n_liq * ᶜuʲs.:($$j)) -
-                    Y.c.sgsʲs.:($$j).n_liq * wdivₕ(ᶜuʲs.:($$j))
+                    split_divₕ(ᶜuʲs.:($$j), Y.c.sgsʲs.:($$j).n_liq) -
+                    Y.c.sgsʲs.:($$j).n_liq * split_divₕ(ᶜuʲs.:($$j), 1)
                 @. Yₜ.c.sgsʲs.:($$j).n_rai -=
-                    wdivₕ(Y.c.sgsʲs.:($$j).n_rai * ᶜuʲs.:($$j)) -
-                    Y.c.sgsʲs.:($$j).n_rai * wdivₕ(ᶜuʲs.:($$j))
+                    split_divₕ(ᶜuʲs.:($$j), Y.c.sgsʲs.:($$j).n_rai) -
+                    Y.c.sgsʲs.:($$j).n_rai * split_divₕ(ᶜuʲs.:($$j), 1)
             end
         end
     end
@@ -187,7 +199,7 @@ This function handles:
   their central advection might be handled elsewhere or implicitly.
 - Vertical advection terms for horizontal and vertical momentum, differing for
   shallow and deep atmosphere approximations, incorporating Coriolis and vorticity effects.
-- Vertical advection of grid-mean TKE (`ρatke⁰`) if `use_prognostic_tke` is true.
+- Vertical advection of grid-mean TKE (`ρtke`) if `use_prognostic_tke` is true.
 
 Arguments:
 - `Yₜ`: The tendency state vector, modified in place.
@@ -199,7 +211,7 @@ Arguments:
 - `t`: Current simulation time (not directly used in calculations).
 
 Modifies `Yₜ.c` (various tracers, `ρe_tot`, `ρq_tot`, `uₕ`), `Yₜ.f.u₃`,
-`Yₜ.f.sgsʲs` (updraft `u₃`), and `Yₜ.c.sgs⁰.ρatke` as applicable.
+`Yₜ.f.sgsʲs` (updraft `u₃`), and `Yₜ.c.ρtke` as applicable.
 """
 NVTX.@annotate function explicit_vertical_advection_tendency!(Yₜ, Y, p, t)
     (; turbconv_model, prescribed_flow) = p.atmos
@@ -215,9 +227,9 @@ NVTX.@annotate function explicit_vertical_advection_tendency!(Yₜ, Y, p, t)
     (; energy_q_tot_upwinding, tracer_upwinding) = p.atmos.numerics
     thermo_params = CAP.thermodynamics_params(p.params)
 
-    ᶜtke⁰ =
+    ᶜtke =
         advect_tke ?
-        (@. lazy(specific(Y.c.sgs⁰.ρatke, Y.c.ρ))) :
+        (@. lazy(specific(Y.c.ρtke, Y.c.ρ))) :
         nothing
     ᶜω³ = p.scratch.ᶜtemp_CT3
     ᶠω¹² = p.scratch.ᶠtemp_CT12
@@ -301,8 +313,8 @@ NVTX.@annotate function explicit_vertical_advection_tendency!(Yₜ, Y, p, t)
     end
 
     if use_prognostic_tke(turbconv_model) # advect_tke triggers allocations
-        vtt = vertical_transport(ᶜρ, ᶠu³, ᶜtke⁰, dt, edmfx_mse_q_tot_upwinding)
-        @. Yₜ.c.sgs⁰.ρatke += vtt
+        vtt = vertical_transport(ᶜρ, ᶠu³, ᶜtke, dt, edmfx_mse_q_tot_upwinding)
+        @. Yₜ.c.ρtke += vtt
     end
 end